refactor: move the resource interface into the agents
todo: implement resource helper methods directly on each resource.
