+++ control-plane/agents/core/src/controller/resources/operations.rs
