+++ control-plane/agents/core/src/controller/scheduling/resources/mod.rs
@@ -11,9 +11,12 @@ use common_lib::types::v0::{
     transport::{Child, ChildUri, Replica},
 };
 
+/// Item for pool scheduling logic.
 #[derive(Debug, Clone)]
 pub(crate) struct PoolItem {
+    /// The node where this pools lives.
     pub(crate) node: NodeWrapper,
+    /// The pool.
     pub(crate) pool: PoolWrapper,
 }
 
@@ -21,11 +24,13 @@ impl PoolItem {
     fn new(node: NodeWrapper, pool: PoolWrapper) -> Self {
         Self { node, pool }
     }
+    /// Collect the item into a pool.
     pub(crate) fn collect(self) -> PoolWrapper {
         self.pool
     }
 }
 
+/// A pool lister.
 pub(crate) struct PoolItemLister {}
 impl PoolItemLister {
     async fn nodes(registry: &Registry) -> Vec<NodeWrapper> {
@@ -37,6 +42,7 @@ impl PoolItemLister {
         }
         raw_nodes
     }
+    /// Get a list of pool items.
     pub(crate) async fn list(registry: &Registry) -> Vec<PoolItem> {
         let pools = Self::nodes(registry)
             .await
@@ -53,6 +59,7 @@ impl PoolItemLister {
     }
 }
 
+/// A replica item used for scheduling.
 #[derive(Debug, Clone)]
 pub(crate) struct ReplicaItem {
     replica_spec: ReplicaSpec,
@@ -64,7 +71,7 @@ pub(crate) struct ReplicaItem {
 }
 
 impl ReplicaItem {
-    /// Create new `Self` from the provided arguments
+    /// Create new `Self` from the provided arguments.
     pub(crate) fn new(
         replica: ReplicaSpec,
         replica_state: Option<&Replica>,
@@ -82,33 +89,33 @@ impl ReplicaItem {
             child_info,
         }
     }
-    /// Get a reference to the replica spec
+    /// Get a reference to the replica spec.
     pub(crate) fn spec(&self) -> &ReplicaSpec {
         &self.replica_spec
     }
-    /// Get a reference to the replica state
+    /// Get a reference to the replica state.
     pub(crate) fn state(&self) -> Option<&Replica> {
         self.replica_state.as_ref()
     }
-    /// Get a reference to the child spec
+    /// Get a reference to the child spec.
     pub(crate) fn uri(&self) -> &Option<ChildUri> {
         &self.child_uri
     }
-    /// Get a reference to the child state
+    /// Get a reference to the child state.
     pub(crate) fn child_state(&self) -> &Option<Child> {
         &self.child_state
     }
-    /// Get a reference to the child spec
+    /// Get a reference to the child spec.
     pub(crate) fn child_spec(&self) -> Option<&NexusChild> {
         self.child_spec.as_ref()
     }
-    /// Get a reference to the child info
+    /// Get a reference to the child info.
     pub(crate) fn child_info(&self) -> Option<&ChildInfo> {
         self.child_info.as_ref()
     }
 }
 
-/// Individual nexus child (replicas) which can be used for nexus creation
+/// Individual nexus child (replicas) which can be used for nexus creation.
 #[derive(Clone)]
 pub(crate) struct ChildItem {
     replica_spec: ReplicaSpec,
@@ -130,30 +137,30 @@ impl std::fmt::Debug for ChildItem {
 }
 
 /// If the nexus is shutdown uncleanly, only one child/replica may be used and it must be healthy
-/// This is to avoid inconsistent data between the healthy replicas
+/// This is to avoid inconsistent data between the healthy replicas.
 #[derive(Debug, Clone)]
 pub(crate) enum HealthyChildItems {
-    /// One with multiple healthy candidates
+    /// One with multiple healthy candidates.
     One(Option<NexusInfo>, Vec<ChildItem>),
-    /// All the healthy replicas can be used
+    /// All the healthy replicas can be used.
     All(Option<NexusInfo>, Vec<ChildItem>),
 }
 impl HealthyChildItems {
-    /// Check if there are no healthy children
+    /// Check if there are no healthy children.
     pub(crate) fn is_empty(&self) -> bool {
         match self {
             HealthyChildItems::One(_, items) => items.is_empty(),
             HealthyChildItems::All(_, items) => items.is_empty(),
         }
     }
-    /// Get a reference to the list of candidates
+    /// Get a reference to the list of candidates.
     pub(crate) fn candidates(&self) -> &Vec<ChildItem> {
         match self {
             HealthyChildItems::One(_, items) => items,
             HealthyChildItems::All(_, items) => items,
         }
     }
-    /// Get a reference to the list of candidates
+    /// Get a reference to the list of candidates.
     pub(crate) fn nexus_info(&self) -> &Option<NexusInfo> {
         match self {
             HealthyChildItems::One(info, _) => info,
@@ -163,7 +170,7 @@ impl HealthyChildItems {
 }
 
 impl ChildItem {
-    /// Create a new `Self` from the replica and the persistent child information
+    /// Create a new `Self` from the replica and the persistent child information.
     pub(crate) fn new(
         replica_spec: &ReplicaSpec,
         replica_state: &Replica,
@@ -177,19 +184,19 @@ impl ChildItem {
             pool_state: pool_state.clone(),
         }
     }
-    /// Get the replica spec
+    /// Get the replica spec.
     pub(crate) fn spec(&self) -> &ReplicaSpec {
         &self.replica_spec
     }
-    /// Get the replica state
+    /// Get the replica state.
     pub(crate) fn state(&self) -> &Replica {
         &self.replica_state
     }
-    /// Get the persisted nexus child information
+    /// Get the persisted nexus child information.
     pub(crate) fn info(&self) -> &Option<ChildInfo> {
         &self.child_info
     }
-    /// Get the pool wrapper
+    /// Get the pool wrapper.
     pub(crate) fn pool(&self) -> &PoolWrapper {
         &self.pool_state
     }
