+++ control-plane/agents/core/src/server.rs
@@ -1,10 +1,20 @@
+//! The Core Agent.
+//! todo: document.
+
+/// The controller logic for all resources.
 pub(crate) mod controller;
+/// The nexus related operations.
 pub(crate) mod nexus;
+/// The node related operations.
 pub(crate) mod node;
+/// The pool related operations.
 pub(crate) mod pool;
+/// The registry which contains all the resources.
 pub(crate) mod registry;
 mod service;
+/// The volume related operations.
 pub(crate) mod volume;
+/// The watch related operations.
 pub(crate) mod watch;
 
 use controller::registry::NumRebuilds;
@@ -14,58 +24,59 @@ use http::Uri;
 use opentelemetry::{trace::TracerProvider, KeyValue};
 use structopt::StructOpt;
 
+/// The Cli arguments for this binary.
 #[derive(Debug, StructOpt)]
 #[structopt(name = utils::package_description!(), version = version_info_str!())]
 pub(crate) struct CliArgs {
     /// The period at which the registry updates its cache of all
-    /// resources from all nodes
+    /// resources from all nodes.
     #[structopt(long, short, default_value = utils::CACHE_POLL_PERIOD)]
     pub(crate) cache_period: humantime::Duration,
 
-    /// The period at which the reconcile loop checks for new work
+    /// The period at which the reconcile loop checks for new work.
     #[structopt(long, default_value = "30s")]
     pub(crate) reconcile_idle_period: humantime::Duration,
 
-    /// The period at which the reconcile loop attempts to do work
+    /// The period at which the reconcile loop attempts to do work.
     #[structopt(long, default_value = "10s")]
     pub(crate) reconcile_period: humantime::Duration,
 
-    /// Deadline for the io-engine instance keep alive registration
+    /// Deadline for the io-engine instance keep alive registration.
     #[structopt(long, short, default_value = "10s")]
     pub(crate) deadline: humantime::Duration,
 
-    /// The Persistent Store URLs to connect to
+    /// The Persistent Store URLs to connect to.
     /// (supports the http/https schema)
     #[structopt(long, short, default_value = "http://localhost:2379")]
     pub(crate) store: String,
 
-    /// The timeout for store operations
+    /// The timeout for store operations.
     #[structopt(long, default_value = utils::STORE_OP_TIMEOUT)]
     pub(crate) store_timeout: humantime::Duration,
 
-    /// The lease lock ttl for the persistent store after which we'll lose the exclusive access
+    /// The lease lock ttl for the persistent store after which we'll lose the exclusive access.
     #[structopt(long, default_value = utils::STORE_LEASE_LOCK_TTL)]
     pub(crate) store_lease_ttl: humantime::Duration,
 
-    /// The timeout for every node connection (gRPC)
+    /// The timeout for every node connection (gRPC).
     #[structopt(long, default_value = utils::DEFAULT_CONN_TIMEOUT)]
     pub(crate) connect_timeout: humantime::Duration,
 
-    /// The default timeout for node request timeouts (gRPC)
+    /// The default timeout for node request timeouts (gRPC).
     #[structopt(long, short, default_value = utils::DEFAULT_REQ_TIMEOUT)]
     pub(crate) request_timeout: humantime::Duration,
 
-    /// Add process service tags to the traces
+    /// Add process service tags to the traces.
     #[structopt(short, long, env = "TRACING_TAGS", value_delimiter=",", parse(try_from_str = utils::tracing_telemetry::parse_key_value))]
     tracing_tags: Vec<KeyValue>,
 
-    /// Don't use minimum timeouts for specific requests
+    /// Don't use minimum timeouts for specific requests.
     #[structopt(long)]
     no_min_timeouts: bool,
-    /// Trace rest requests to the Jaeger endpoint agent
+    /// Trace rest requests to the Jaeger endpoint agent.
     #[structopt(long, short)]
     jaeger: Option<String>,
-    /// The GRPC Server URLs to connect to
+    /// The GRPC Server URLs to connect to.
     /// (supports the http/https schema)
     #[structopt(long, short, default_value = DEFAULT_GRPC_SERVER_ADDR)]
     pub(crate) grpc_server_addr: Uri,
